chore: fix typings
