+++ web/apps/benchmarks-and-reports/src/app/(benchmarks-and-reports)/benchmarks/_components/charts.tsx
@@ -4,160 +4,22 @@ import { Button } from "@risc0/ui/button";
 import { Separator } from "@risc0/ui/separator";
 import { Skeleton } from "@risc0/ui/skeleton";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@risc0/ui/tabs";
-import { Chart } from "chart.js"; // stay on 2.9.4, much faster
-import truncate from "lodash-es/truncate";
 import { DownloadIcon } from "lucide-react";
 import Script from "next/script";
 import { useEffect, useState } from "react";
+import { type FormattedDataSetEntry, collectBenchesPerTestCase } from "../_utils/collect-benches-per-test-case";
+import { renderBenchSet } from "../_utils/render-bench-set";
 import { ChartsList } from "./charts-list";
-
-const toolColors = {
-  cargo: "#020077",
-  go: "#00add8",
-  benchmarkjs: "#f1e05a",
-  benchmarkluau: "#000080",
-  pytest: "#3572a5",
-  googlecpp: "#f34b7d",
-  catch2: "#f34b7d",
-  julia: "#a270ba",
-  jmh: "#b07219",
-  benchmarkdotnet: "#178600",
-  customBiggerIsBetter: "#38ff38",
-  customSmallerIsBetter: "#ff3838",
-  _: "#333333",
-};
-
-function renderGraph({ parent, benchName, platformName, dataset }) {
-  const canvas = document.createElement("canvas");
-  canvas.className = "benchmark-chart";
-  canvas.id = `${platformName}-${benchName}`;
-  parent.appendChild(canvas);
-
-  const color = toolColors[dataset.length > 0 ? dataset[0].tool : "_"];
-
-  const data = {
-    labels: dataset.map((d) => d.commit.id.slice(0, 7)),
-    datasets: [
-      {
-        label: benchName,
-        data: dataset.map((d) => d.bench.value),
-        borderColor: color,
-        borderWidth: 1,
-        fill: true,
-        backgroundColor: `${color}50`, // Add alpha for #rrggbbaa
-      },
-    ],
-  };
-
-  const options = {
-    animation: {
-      duration: 0, // general animation time
-    },
-    hover: {
-      animationDuration: 0, // duration of animations when hovering an item
-    },
-    responsiveAnimationDuration: 0, // animation duration after a resize
-    aspectRatio: 4,
-    scales: {
-      xAxes: [
-        {
-          scaleLabel: {
-            display: true,
-            labelString: "commit",
-          },
-        },
-      ],
-      yAxes: [
-        {
-          scaleLabel: {
-            display: true,
-            labelString: dataset.length > 0 ? dataset[0].bench.unit : "",
-          },
-          ticks: {
-            beginAtZero: true,
-          },
-        },
-      ],
-    },
-    onClick: (_mouseEvent, activeElems) => {
-      if (activeElems.length === 0) {
-        return;
-      }
-
-      const index = activeElems[0]._index;
-      const url = dataset[index].commit.url;
-
-      window.open(url, "_blank");
-    },
-    onHover: (event, chartElement) => {
-      event.target.style.cursor = chartElement[0] ? "pointer" : "default";
-    },
-    tooltips: {
-      backgroundColor: "rgba(0, 0, 0, 1)",
-      callbacks: {
-        afterTitle: (items) => {
-          const { index } = items[0];
-          const data = dataset[index];
-          return `\n${truncate(data.commit.message, {
-            length: 140,
-            omission: "…",
-          })}\n\n${data.commit.timestamp} committed by @${data.commit.committer.username}\n`;
-        },
-        label: (item) => {
-          let label = item.value;
-          const { range, unit } = dataset[item.index].bench;
-          label += ` ${unit}`;
-          if (range) {
-            label += ` (${range})`;
-          }
-          return label;
-        },
-        afterLabel: (item) => {
-          const { extra } = dataset[item.index].bench;
-          return extra ? `\n${extra}` : "";
-        },
-      },
-    },
-  };
-
-  new Chart(canvas, {
-    type: "line",
-    data,
-    options,
-  });
-}
-
-function renderBenchSet({ platformName, benchSet, main }) {
-  const graphsElem = document.createElement("div");
-  graphsElem.className = "mt-6 flex flex-row flex-wrap gap-10 dark:invert [&>*]:tracking-normal";
-  main.appendChild(graphsElem);
-
-  for (const [benchName, benches] of benchSet.entries()) {
-    renderGraph({ parent: graphsElem, benchName, dataset: benches, platformName });
-  }
-}
-
-function collectBenchesPerTestCase(entries) {
-  const map = new Map();
-
-  for (const entry of entries) {
-    const { commit, date, tool, benches } = entry;
-    for (const bench of benches) {
-      const result = { commit, date, tool, bench };
-      const arr = map.get(bench.name);
-      if (arr === undefined) {
-        map.set(bench.name, [result]);
-      } else {
-        arr.push(result);
-      }
-    }
-  }
-  return map;
-}
 
 export function Charts() {
   const [lastUpdate, setLastUpdate] = useState<string>("");
-  const [benchSet, setBenchSet] = useState<any>();
+  const [benchSet, setBenchSet] =
+    useState<
+      {
+        name: string;
+        dataSet: Map<string, FormattedDataSetEntry[]>;
+      }[]
+    >();
   const [selectedPlatform, setSelectedPlatform] = useState<string>();
   const [names, setNames] = useState<string[]>();
   const [ready, setReady] = useState<boolean>(false);
@@ -201,7 +63,8 @@ export function Charts() {
     setBenchSet(dataset);
 
     for (const { name, dataSet } of dataset) {
-      renderBenchSet({ platformName: name, benchSet: dataSet, main: document.getElementById(`chart-${name}`) });
+      // biome-ignore lint/style/noNonNullAssertion: ignore
+      renderBenchSet({ platformName: name, benchSet: dataSet, main: document.getElementById(`chart-${name}`)! });
     }
   }, [names]);
 
@@ -251,7 +114,7 @@ export function Charts() {
 
           <div className="mt-4 flex flex-row gap-8">
             <div className="sticky top-6 hidden min-w-[280px] self-start lg:block">
-              <ChartsList charts={benchSet} selectedPlatform={selectedPlatform} />
+              {benchSet && selectedPlatform && <ChartsList charts={benchSet} selectedPlatform={selectedPlatform} />}
             </div>
 
             <div className="w-full">
